requirements.txt
Flask==0.12.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.2
Flask-Testing==0.7.1
factory-boy==2.9.2
gunicorn==19.7.1
coverage==4.4.2
coverage==4.4.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Bcrypt==0.7.1
marshmallow==2.15.0
PyJWT==1.5.3
freezegun==0.3.9
